+++ src/extension/api.ts
@@ -96,13 +96,13 @@ export class API extends EventEmitter<RooCodeEvents> implements RooCodeAPI {
 		configuration,
 		text,
 		images,
-		files,
+		filePaths,
 		newTab,
 	}: {
 		configuration: RooCodeSettings
 		text?: string
 		images?: string[]
-		files?: string[]
+		filePaths?: string[]
 		newTab?: boolean
 	}) {
 		let provider: ClineProvider
@@ -132,9 +132,9 @@ export class API extends EventEmitter<RooCodeEvents> implements RooCodeAPI {
 		await provider.removeClineFromStack()
 		await provider.postStateToWebview()
 		await provider.postMessageToWebview({ type: "action", action: "chatButtonClicked" })
-		await provider.postMessageToWebview({ type: "invoke", invoke: "newChat", text, images, filePaths: files })
+		await provider.postMessageToWebview({ type: "invoke", invoke: "newChat", text, images, filePaths })
 
-		const { taskId } = await provider.initClineWithTask(text, images, files, undefined, {
+		const { taskId } = await provider.initClineWithTask(text, images, filePaths, undefined, {
 			consecutiveMistakeLimit: Number.MAX_SAFE_INTEGER,
 		})
