+++ src/integrations/misc/extract-text.ts
@@ -411,35 +411,3 @@ function processLineWithCarriageReturns(
 
 	return curLine
 }
-
-// kilocode_change start
-/**
- * Helper function used to load file(s) and format them into a string
- */
-export async function processFilesIntoText(files: string[]): Promise<string> {
-	const fileContentsPromises = files.map(async (filePath) => {
-		try {
-			const content = await extractTextFromFile(filePath)
-			// Normalize path separators to forward slashes
-			const normalizedPath = filePath.split(path.sep).join("/")
-			return `<file_content path="${normalizedPath}">\n${content}\n</file_content>`
-		} catch (error) {
-			console.error(`Error processing file ${filePath}:`, error)
-			const normalizedPath = filePath.split(path.sep).join("/")
-			const errorMessage = error instanceof Error ? error.message : String(error)
-			return `<file_content path="${normalizedPath}">\nError fetching content: ${errorMessage}\n</file_content>`
-		}
-	})
-
-	const fileContents = await Promise.all(fileContentsPromises)
-
-	const validFileContents = fileContents.filter((content) => content !== null).join("\n\n")
-
-	if (validFileContents) {
-		return `Files attached by the user:\n\n${validFileContents}`
-	}
-
-	// returns empty string if no files were loaded properly
-	return ""
-}
-// kilocode_change end
