+++ src/package.json
@@ -458,7 +458,6 @@
 		"google-auth-library": "^9.15.1",
 		"i18next": "^24.2.2",
 		"ignore": "^7.0.3",
-		"image-size": "^2.0.2",
 		"isbinaryfile": "^5.0.2",
 		"lodash.debounce": "^4.0.8",
 		"lru-cache": "^11.1.0",
